[Feat] 선택요구사항 8. 레벨업 기능 추가

+++ TextRPG/TextRPG/Interface.cs
@@ -11,8 +11,8 @@
     {
         string name { get; set; }
         CreatureType classType { get; set; }
-        int attackPower { get; set; }
-        int defensePower { get; set; }
-        int health { get; set; }
+        float attackPower { get; set; }
+        float defensePower { get; set; }
+        float health { get; set; }
     }
 }
